+++ appinfo/routes.php
@@ -91,6 +91,7 @@ return [
 		['name' => 'datasource#index', 'url' => '/datasource', 'verb' => 'GET'],
 		['name' => 'datasource#own', 'url' => '/datasource/own', 'verb' => 'GET'],
 		['name' => 'datasource#registered', 'url' => '/datasource/registered', 'verb' => 'GET'],
+		['name' => 'datasource#spreadsheetSheets', 'url' => '/datasource/spreadsheet/sheets', 'verb' => 'GET'],
 		['name' => 'datasource#indexFiltered', 'url' => '/datasource/{datasourceType}', 'verb' => 'GET'],
 
 		// Share