+++ tests/Stubs/OCP/Files/File.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * SPDX-FileCopyrightText: 2026 Marcel Scherello
+ * SPDX-License-Identifier: AGPL-3.0-or-later
+ */
+
+namespace OCP\Files;
+
+interface File {
+	public function getStorage();
+	public function getInternalPath();
+}